Refactor DependencyClassLoader to separate to SelfContainedJarAwareURLClassLoader
It prepares for PluginClassLoader to be able to load from JAR resources self-contained in the
Embulk JAR file. PluginClassLoader is going to extend SelfContainedJarAwareURLClassLoader, too.
